feat: mvp inAppMessaging
