+++ packages/amplify-category-notifications/src/commands/notifications/status.ts
@@ -4,11 +4,15 @@ import {
 import chalk from 'chalk';
 import { IChannelAvailability, INotificationsConfigStatus } from '../../notifications-api-types';
 import { NotificationsDB } from '../../notifications-backend-cfg-api';
+import { NotificationsMeta } from '../../notifications-meta-api';
 
 const viewStyles = {
   enabled: chalk.bold.green,
   disabled: chalk.bold.red,
   notDeployed: chalk.dim,
+  url: chalk.bold.yellow,
+  underline: chalk.blue.underline,
+  appName: chalk.bold.yellowBright,
 };
 
 const getDeployedStyledStatus = (deployedChannel: string, deployedChannels: IChannelAvailability): string => {
@@ -18,6 +22,14 @@ const getDeployedStyledStatus = (deployedChannel: string, deployedChannels: ICha
   return (deployedChannels.disabledChannels.includes(deployedChannel))
     ? viewStyles.disabled('Disabled')
     : viewStyles.notDeployed('Not Deployed');
+};
+
+const viewNotificationsAppURL = async (context: $TSContext, appName: string): Promise<void> => {
+  const meta = await NotificationsMeta.getNotificationsAppMeta(context.exeInfo.amplifyMeta, appName);
+  if (meta?.Id) {
+    const consoleUrl = `https://${meta.Region}.console.aws.amazon.com/pinpoint/home/?region=${meta.Region}#/apps/${meta.Id}/notifications`;
+    context.print.info(`\nPinpoint App: ${viewStyles.underline(viewStyles.url(consoleUrl))}`);
+  }
 };
 
 const viewDisplayChannelAvailability = async (context: $TSContext, backend:INotificationsConfigStatus): Promise<void> => {
@@ -32,7 +44,7 @@ const viewDisplayChannelAvailability = async (context: $TSContext, backend:INoti
 };
 
 const viewDisplayNotificationsResourceInfo = async (context: $TSContext, backend:INotificationsConfigStatus):Promise<void> => {
-  context.print.info(`Application Name: ${backend.local.config.serviceName} (${backend.local.config.service})`);
+  context.print.info(`\n\nApplication : ${viewStyles.appName(backend.local.config.serviceName)} (${backend.local.config.service})`);
 };
 
 /**
@@ -43,6 +55,7 @@ export const run = async (context:$TSContext):Promise<void> => {
   const backend: INotificationsConfigStatus|undefined = await NotificationsDB.ChannelAPI.getNotificationConfigStatus(context);
   if (backend) {
     await viewDisplayNotificationsResourceInfo(context, backend);
+    await viewNotificationsAppURL(context, backend.local.config.serviceName);
     await viewDisplayChannelAvailability(context, backend);
   }
 };
