+++ packages/amplify-category-notifications/src/notifications-backend-cfg-api.ts
@@ -11,163 +11,13 @@ import {
   $TSAny, pathManager, stateManager, AmplifySupportedService, AmplifyCategories, $TSContext,
 } from 'amplify-cli-core';
 import {
-  ChannelAction, ChannelConfigDeploymentType, INotificationsConfigStatus, IChannelAPIResponse, IChannelAvailability,
+  ChannelAction, IChannelAPIResponse,
 } from './notifications-api-types';
-import { INotificationsChannelBackendConfig, INotificationsResourceBackendConfig, INotificationsResourceBackendConfigValue } from './notifications-backend-config-types';
+import { ChannelAPI } from './notifications-backend-cfg-channel-api';
+import { INotificationsResourceBackendConfig, INotificationsResourceBackendConfigValue } from './notifications-backend-config-types';
 import { NotificationsMeta } from './notifications-meta-api';
 
 const PROVIDER_NAME = 'awscloudformation';
-
-/**
- * API to update Notification Channel config state
- * All functions are idempotent (no side effects)
- */
-export class ChannelAPI {
-  /**
-   * Channel names
-   */
-  public static ChannelType: Record<string, string> = {
-    APNS: 'APNS',
-    FCM: 'FCM',
-    Email: 'Email',
-    SMS: 'SMS',
-    InAppMessaging: 'InAppMessaging',
-  }
-
-  /**
-   * Map of channel-type to channel-code
-   */
-   public static channelWorkers: Record<string, string> = {
-     [ChannelAPI.ChannelType.APNS]: './channel-APNS',
-     [ChannelAPI.ChannelType.FCM]: './channel-FCM',
-     [ChannelAPI.ChannelType.Email]: './channel-Email',
-     [ChannelAPI.ChannelType.SMS]: './channel-SMS',
-     [ChannelAPI.ChannelType.InAppMessaging]: './channel-InAppMessaging',
-   };
-
-  public static isValidChannel = (channelName: string| undefined): boolean => (channelName !== undefined
-    && channelName in ChannelAPI.ChannelType);
-
-    /**
-     * For a given notifications resource get local and deployed channel availability
-     * @param backendResourceConfig notifications resource info from the backend config
-     * @returns enabled and disabled channels
-     */
-    public static getChannelAvailability = async (backendResourceConfig:INotificationsResourceBackendConfig)
-    : Promise<IChannelAvailability> => {
-      const availableChannels = NotificationsDB.ChannelAPI.getAvailableChannels();
-      const enabledChannels = (await NotificationsDB.getEnabledChannelsFromBackendConfig(backendResourceConfig)) || [];
-      const disabledChannels = (await NotificationsDB.getDisabledChannelsFromBackendConfig(availableChannels, enabledChannels)) || [];
-      const backend : IChannelAvailability = {
-        enabledChannels,
-        disabledChannels,
-      };
-      return backend;
-    };
-
-    /**
-     * Get notifications resource localBackend, deployedBackend and channel availability
-     * most useful in displaying status.
-     * @param context amplify cli context
-     * @returns backendConfig and channel availability for notifications
-     */
-    public static getNotificationConfigStatus = async (context:$TSContext): Promise<INotificationsConfigStatus|undefined> => {
-      const notificationConfig = await NotificationsDB.getNotificationsAppConfig(context.exeInfo.backendConfig);
-
-      // no Notifications resource
-      if (!notificationConfig) {
-        return undefined;
-      }
-      let appInitialized = true;
-      let deployedBackendConfig: $TSAny;
-      try {
-        deployedBackendConfig = (stateManager.getCurrentBackendConfig()) || undefined;
-      } catch (e) {
-        appInitialized = false;
-        deployedBackendConfig = undefined;
-      } // this will fail on iniEnv;
-      const deployedNotificationConfig = await NotificationsDB.getNotificationsAppConfig(deployedBackendConfig);
-      const emptyChannels = { enabledChannels: [], disabledChannels: [] };
-      return {
-        local: {
-          config: notificationConfig,
-          channels: (notificationConfig) ? await ChannelAPI.getChannelAvailability(notificationConfig) : emptyChannels,
-        },
-        deployed: {
-          config: deployedNotificationConfig,
-          channels: (deployedNotificationConfig) ? await ChannelAPI.getChannelAvailability(deployedNotificationConfig) : emptyChannels,
-        },
-        appInitialized,
-      } as INotificationsConfigStatus;
-    };
-
-    /**
-   * Returns true if resource is deployed only during amplify push
-   * @param validChannelName - a valid channel name
-   * @returns true if channel deployment is handled through amplify push
-   */
-    public static isChannelDeploymentDeferred = (validChannelName: string): boolean => (
-      ChannelAPI.getChannelDeploymentType(validChannelName) === ChannelConfigDeploymentType.DEFERRED
-    )
-
-    /**
- * Check if notification channel has been added to the backend-config
- * @param resourceBackendConfig - Backend config for the given pinpoint resource from backend-config.json
- * @param channel - Notification channel to be checked for.
- * @returns true if channel is enabled in backend-config
- */
-  public static isNotificationChannelEnabledInBackendConfig = (resourceBackendConfig: INotificationsResourceBackendConfig,
-    channel: string):boolean => resourceBackendConfig.channels && resourceBackendConfig.channels.includes(channel);
-
-  public static getNotificationChannelEnabledInBackendConfig = (resourceBackendConfig: INotificationsResourceBackendConfig)
-  :Array<string> => resourceBackendConfig.channels
-
-  /**
- * Get all available notification channels
- */
- public static getAvailableChannels = ():Array<string> => Object.keys(ChannelAPI.channelWorkers);
-
- public static getChannelDeploymentType =
- (channelName: string): ChannelConfigDeploymentType => ((channelName === ChannelAPI.ChannelType.InAppMessaging)
-   ? ChannelConfigDeploymentType.DEFERRED
-   : ChannelConfigDeploymentType.INLINE)
-
-  public static enableNotificationsChannel = (notificationsConfig: INotificationsResourceBackendConfig,
-    validChannelName: string, channelConfig?: INotificationsChannelBackendConfig):INotificationsResourceBackendConfig => {
-    if (notificationsConfig.channels && !notificationsConfig.channels.includes(validChannelName)) {
-      notificationsConfig.channels.push(validChannelName);
-      if (channelConfig) {
-        notificationsConfig.channelConfig[validChannelName] = channelConfig;
-      }
-      return notificationsConfig;
-    }
-    throw new Error(`EnableNotificationsChannel Failed: Invalid notificationsConfig: ${JSON.stringify(notificationsConfig, null, 2)}`);
-  }
-
-  public static disableNotificationsChannel = (notificationsConfig: INotificationsResourceBackendConfig,
-    validChannelName: string):INotificationsResourceBackendConfig => {
-    if (notificationsConfig.channels && !notificationsConfig.channels.includes(validChannelName)) {
-      notificationsConfig.channels = notificationsConfig.channels.filter(channelName => channelName !== validChannelName);
-      if (notificationsConfig.channelConfig && validChannelName in notificationsConfig.channelConfig) {
-        delete notificationsConfig.channelConfig[validChannelName];
-      }
-      return notificationsConfig;
-    }
-    throw new Error(`EnableNotificationsChannel Failed: Invalid notificationsConfig: ${JSON.stringify(notificationsConfig, null, 2)}`);
-  }
-
-  public static updateNotificationsChannelConfig = (notificationsConfig: INotificationsResourceBackendConfig,
-    validChannelName: string, channelConfig: INotificationsChannelBackendConfig):INotificationsResourceBackendConfig => {
-    if (notificationsConfig.channels && !notificationsConfig.channels.includes(validChannelName)) {
-      notificationsConfig.channels = notificationsConfig.channels.filter(channelName => channelName !== validChannelName);
-      if (notificationsConfig.channelConfig) {
-        notificationsConfig.channelConfig[validChannelName] = channelConfig;
-      }
-      return notificationsConfig;
-    }
-    throw new Error(`UpdateNotificationsChannelConfig Failed: Invalid notificationsConfig: ${JSON.stringify(notificationsConfig, null, 2)}`);
-  }
-}
 
 /**
  * Persistent state management class for Notifications.
@@ -207,10 +57,10 @@ export class NotificationsDB {
       }
       context.exeInfo.backendConfig[AmplifyCategories.NOTIFICATIONS][notificationConfig.serviceName] = notificationConfig;
     }
-    console.log('SACPCDEBUG: UpdateChannelAPIResponse : context.exeInfo.backendConfig : ',
-      JSON.stringify(context.exeInfo.backendConfig[AmplifyCategories.NOTIFICATIONS], null, 2));
-    console.log('SACPCDEBUG: UpdateChannelAPIResponse : context.exeInfo.amplifyMeta.notifications : ',
-      JSON.stringify(context.exeInfo.amplifyMeta[AmplifyCategories.NOTIFICATIONS], null, 2));
+    // console.log('SACPCDEBUG: UpdateChannelAPIResponse : context.exeInfo.backendConfig : ',
+    //  JSON.stringify(context.exeInfo.backendConfig[AmplifyCategories.NOTIFICATIONS], null, 2));
+    // console.log('SACPCDEBUG: UpdateChannelAPIResponse : context.exeInfo.amplifyMeta.notifications : ',
+    //  JSON.stringify(context.exeInfo.amplifyMeta[AmplifyCategories.NOTIFICATIONS], null, 2));
     return context;
   }
